sistema-certificados/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
Jinja2==3.0.3
itsdangerous==2.0.1
psycopg2-binary==2.9.3
gunicorn==20.1.0